third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/background-repeat-expected.txt

This is a testharness.js-based test.
FAIL Can set 'background-position' to CSS-wide keywords assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to var() references assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to the 'repeat-x' keyword Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'repeat-y' keyword Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'repeat' keyword Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'space' keyword Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'round' keyword Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'no-repeat' keyword Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting 'background-position' to a length throws TypeError
PASS Setting 'background-position' to a percent throws TypeError
PASS Setting 'background-position' to a time throws TypeError
PASS Setting 'background-position' to an angle throws TypeError
PASS Setting 'background-position' to a flexible length throws TypeError
PASS Setting 'background-position' to a number throws TypeError
PASS Setting 'background-position' to a position throws TypeError
PASS Setting 'background-position' to a URL throws TypeError
PASS Setting 'background-position' to a transform throws TypeError
FAIL 'background-position' does not supported 'space repeat' assert_not_equals: Unsupported value must not be null got disallowed value null
Harness: the test ran to completion.
